Splittig out Synthetix test logic into appropriate worker contract tests
